requirements.txt
numpy==1.26.4
openai-whisper==20240930
Wave==0.0.2
openai==1.66.3  # Latest release as of 2025-03-17
customtkinter==5.2.2
tkinter-tooltip  # Needed to show tooltips for ctk components
pyperclip
PyYAML
PyAudio
pyaudiowpatch
soundfile
gtts
# Playsound version 1.3 has issues in playing back audio files
# in case of continuous play back of files in quick succession
playsound==1.2.2
deepgram-sdk==3.2.5
# Use 117 to build for CPU only
# --extra-index-url https://download.pytorch.org/whl/cu117
#--extra-index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://download.pytorch.org/whl/cu121
torch
bandit==1.7.8
SqlAlchemy==2.0.29
setuptools
appdirs
pillow==11.1.0
wordcloud==1.9.4
ffmpeg-python

# Monitoring system dependencies
aiohttp==3.11.11
aiohttp-sse==2.2.0
aiohttp-cors==0.7.0
requests==2.32.3
